fix: compile with dunmaic linking pods
