+++ example/pubspec.yaml
@@ -18,11 +18,6 @@ dependencies:
     sdk: flutter
 
   yuno_flutter_plugin:
-    # When depending on this package from a real application you should use:
-    #   yuno_flutter_plugin: ^x.y.z
-    # See https://dart.dev/tools/pub/dependencies#version-constraints
-    # The example app is bundled with the plugin so we use a path dependency on
-    # the parent directory to use the current plugin's version.
     path: ../
 
   # The following adds the Cupertino Icons font to your application.
