+++ ios/yuno_flutter_plugin.podspec
@@ -16,8 +16,23 @@ Flutter plugin project for Yuno SDK.
   s.source_files = 'Classes/**/*'
   s.dependency 'Flutter'
   s.dependency 'YunoSDK','1.13.0'
+  s.static_framework = true
   s.platform = :ios, '13.0'
+  s.ios.deployment_target = '13.0'
+  s.requires_arc = true
+  s.resource_bundles = {
+      'Resources' => [
+          'YunoSDK.xcframework/Assets/**/*.xcassets',
+          'YunoSDK.xcframework/Assets/Localization/*.lproj',
+          'YunoSDK.xcframework/Assets/**/*.{storyboard,xib,json,ttf}'
+      ]
+  }
+  s.xcconfig = { 'SWIFT_INCLUDE_PATHS' => ['${PODS_XCFRAMEWORKS_BUILD_DIR}/YunoSDK'] }
   # Flutter.framework does not contain a i386 slice.
-  s.pod_target_xcconfig = { 'DEFINES_MODULE' => 'YES', 'EXCLUDED_ARCHS[sdk=iphonesimulator*]' => 'i386, arm64' }
-  s.swift_version = '5.0'
+  s.pod_target_xcconfig = { 
+    'DEFINES_MODULE' => 'YES', 
+    'EXCLUDED_ARCHS[sdk=iphonesimulator*]' => 'i386, arm64', 
+    'BUILD_LIBRARY_FOR_DISTRIBUTION' => 'YES' 
+  }
+  s.swift_version = '5.4'
 end
